Add check existence

+++ exec-benchmark.sh
@@ -1,16 +1,27 @@
 #!/bin/bash
 
 pushd benchmark/date/
-wget http://aozora-word.hahasoha.net/aozora_word_list_utf8.csv.gz
-gunzip -d aozora_word_list_utf8.csv.gz
-wget http://jusyo.jp/downloads/new/csv/csv_zenkoku.zip
-unzip csv_zenkoku.zip
-wget http://dumps.wikimedia.org/jawiki/latest/jawiki-latest-all-titles.gz
-gunzip -d jawiki-latest-all-titles.gz
-sed s/\"/\"\"/g jawiki-latest-all-titles | sed s/^/\"/g | sed s/$/\"/g > jawiki-latest-all-titles.tmp
+
+if [ ! -f ./aozora_word_list_utf8.csv.gz ]; then
+  wget http://aozora-word.hahasoha.net/aozora_word_list_utf8.csv.gz
+  gunzip -d aozora_word_list_utf8.csv.gz
+fi
+
+if [ ! -f ./csv_zenkoku.zip ]; then
+  wget http://jusyo.jp/downloads/new/csv/csv_zenkoku.zip
+  unzip csv_zenkoku.zip
+fi
+
+if [ ! -f ./jawiki-latest-all-titles.gz ]; then
+  wget http://dumps.wikimedia.org/jawiki/latest/jawiki-latest-all-titles.gz
+  gunzip -d jawiki-latest-all-titles.gz
+  sed s/\"/\"\"/g jawiki-latest-all-titles | sed s/^/\"/g | sed s/$/\"/g > jawiki-latest-all-titles.tmp
+fi
+
 popd
 
 sudo docker-compose up -d
+
 sleep 3
 psql -h localhost -U PGroonga --command 'CREATE DATABASE benchmark'
 psql -h localhost -U PGroonga -d benchmark -f benchmark/aozora.sql
